cciot/mqtt_publisher/requirements.txt
awsiotsdk
paho-mqtt

